// frontend/src/types/doctypes.ts
interface DocType {
  name: string
  creation: string
  modified: string
  owner: string
  modified_by: string
}
// ...
export interface GPPinnedProject extends DocType {
  /** Project: Link (GP Project) */
  project: string
  /** Team: Link (GP Team) */
  team?: string
  /** User: Link (User) */
  user: string
  /** Order: Int */
  order?: number
}